refactor: bundle the webview chat script from TypeScript modules
Replace the hand-written media/chat.js bundle with the src/webview/chat
module graph and point esbuild at its TypeScript entry, dropping the
now-unused media/chat.js lint globals.
