+++ src/webview/chat/markdown-actions.ts
@@ -1,0 +1,12 @@
+import type { MarkdownActions } from '../markdown.js'
+import { post, t } from './context.js'
+import { copyText } from './utils.js'
+
+export const markdownActions: MarkdownActions = {
+  openExternal: (url) => post('openExternal', { url }),
+  openFile: (reference) => post('openFile', reference as unknown as Record<string, unknown>),
+  copyCode: (code) => copyText(code),
+  defaultCodeLanguage: t('code'),
+  copyLabel: t('copy'),
+  copyCodeLabel: (language) => t('copyCode', { language }),
+}
